+++ blomsterLandet/src/App.js
@@ -6,7 +6,7 @@ import { Header } from './components/common/index';
 import * as firebaseHandler from './firebaseHandler';
 
 import NotificationScreen from './screens/NotificationScreen';
-import MyPlantScreen from './screens/MyPlantScreen';
+import MyPlantScreen from './screens/myplants/MyPlantScreen';
 import ProfileScreen from './screens/ProfileScreen';
 import ShopScreen from './screens/shop/ShopScreen';
 import DevScreen from './screens/DevScreen';
@@ -45,7 +45,13 @@ export const TabNavigator = createBottomTabNavigator({
                     default:
                         break;
                 }
-                return <Icon name={iconName} fontSize={30} flex={1} size={25} color={focused ? '#adc22d' : '#000'} />;
+                return (<Icon 
+                    name={iconName}
+                    fontSize={30} 
+                    flex={1} 
+                    size={25} 
+                    color={focused ? '#adc22d' : '#000'} 
+                />);
             }
         })
     }
